+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v25.47
 
+## v25.xx
+
+### Features
+- New output notifications. (#117 v25.50-alpha) 
+
 ### Features
 - Jinja template support for error notifications. (#112 v25.42-alpha2) 
 - Tenant configuration support for email. To, Cc , Bcc and subject are supported. (#108 v25.42-alpha2)