+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Features
 - New output push notifications. (#117 v25.50-alpha)
--  Push notification supports tenant configuration. (#121 v26.02-alpha)
+- Push notification supports tenant configuration. (#121 v26.02-alpha)
 ## v25.47
 
 ### Features